refactor: Fix designated issues from feedback
