+++ src/App.js
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
-import Footer from "./footer/Footer.js";
-import NavBar from "./nav-bar/NavBar.js";
-import getUserData from "./userData.js";
+import Footer from "./components/Footer/Footer.js";
+import NavBar from "./components/NavBar/NavBar.js";
+import getUserData from "./getUserData.js";
 import "./global.css";
 
 const App = () => {
@@ -15,6 +15,7 @@ const App = () => {
         setProfile({ email, profileImageSource });
       } catch (error) {
         console.error("Error fetching user data:", error);
+        throw error;
       }
     };
     fetchData();
