+++ src/getUserData.js
@@ -10,6 +10,7 @@ const getUserData = async () => {
     return userData;
   } catch (error) {
     console.error("Error fetching user data:", error);
+    throw error;
   }
 };
